+++ tests/test/react/entity_world_reactor.rs
@@ -113,7 +113,7 @@ impl EntityWorldReactor for FullDataReactorMutable
         SystemCommandCallback::new(
             move |mut data: EntityLocal<Self>, event: EntityEvent<usize>|
             {
-                let (event_entity, event_data) = event.read().unwrap();
+                let (event_entity, event_data) = event.read();
                 let (entity, entity_data) = data.get_mut();
 
                 assert_eq!(event_entity, entity);